+++ app/index.js
@@ -209,6 +209,8 @@ module.exports = yeoman.Base.extend({
             this.prompts = props;
             this.props = {};
 
+            this.props.githubUser = (props.githubUser) ? props.githubUser : '{Github User}';
+
             this.props.project = {
                 name: (props.projectName) ? _s.clean(props.projectName) : '{Project Name}',
                 cleanName: (props.projectName) ? _s.clean(props.projectName) : 'project-name',
@@ -220,7 +222,8 @@ module.exports = yeoman.Base.extend({
                 joinedKeywords: props.keywords && props.keywords.join()
             };
 
-            this.props.githubUser = (props.githubUser) ? props.githubUser : '{Github User}';
+            this.props.project.repository = 'https://github.com/' + this.props.githubUser + '/' + this.props.project.sanitizeName + '.git';
+
 
             this.props.author = {
                 name: _s.clean(props.authorName),
@@ -304,6 +307,7 @@ module.exports = yeoman.Base.extend({
         packageJson.keywords = this.props.project.keywords;
         packageJson.author.name = this.props.author.name;
         packageJson.author.url = this.props.author.homepage;
+        packageJson.repository.url = this.props.project.repository;
 
         (this.props.preprocessor.name === 'sass') && (packageJson.devDependencies['gulp-sass'] = '2.3.2');
         (this.props.preprocessor.name === 'stylus') && (packageJson.devDependencies['gulp-stylus'] = '2.5.0');
@@ -569,8 +573,11 @@ module.exports = yeoman.Base.extend({
             bowerJson.name = this.props.project.sanitizeName;
             bowerJson.description = this.props.project.description;
             bowerJson.homepage = this.props.project.homepage;
-            bowerJson.author.name = this.props.author.name;
-            bowerJson.author.homepage = this.props.author.homepage;
+            bowerJson.keywords = this.props.project.keywords;
+            bowerJson.authors[0].name = this.props.author.name;
+            bowerJson.authors[0].homepage = this.props.author.homepage;
+            bowerJson.repository.url = this.props.project.repository;
+            bowerJson.ignore.push('!' + this.props.folder.dest + '/**/*');
 
             this.fs.writeJSON(this.destinationPath('bower.json'), bowerJson);
         }